### 2022-12-03 #4990 @editorialbot
[PRE REVIEW]: Biosiglive: an Open Source Python Package for Real-time Biosignal Processing.
**Submitting author:** @aceglia (<a href="http://orcid.org/0000-0002-7854-9410">Amedeo Ceglia</a>)
**Repository:** https://github.com/pyomeca/biosiglive
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
