+++ docs/paper.md
@@ -114,9 +114,11 @@ while True:
     save({"raw_emg": raw_emg, "process_emg":emg_proc[:, -1]}, "emg.bio")
 ```
 
-The live plot is shown in the following figure.
-[Real-time display of processed ad raw EMG signals for a 5-second window.
+The live plot is shown in the following figure: 
+
+![Real-time display of processed and raw EMG signals for a 5-second window.
 \label{fig:emg_plot}](EMG_plot.png)
+*Real-time display of processed (left) ad raw (right) EMG signals for a 5-second window.*
 
 
 # Research Projects Using `biosiglive`
