### 2022-12-04 #4990 @arfon
@aceglia – this PR should fix your paper compilation: https://github.com/pyomeca/biosiglive/pull/3
