### 2022-12-05 #4990 @Kevin-Mattheus-Moerman
@aceglia thanks for this submission. A quick comment, check your title, perhaps (correct me if I'm wrong) it should be written:

`Biosiglive: an Open Source Python Package for Real-Rime Biosignal Processing.`

i.e. I removed the plural on `sources` and `biosignals`, and I added capital letters throughout (as it previously was only used for Open Source, alternatively remove the other capital letters too).
