### 2022-12-05 #4990 @aceglia
@Kevin-Mattheus-Moerman You are completely right, the title should be :
```Biosiglive: an Open-Source Python Package for Real-Time Biosignal Processing```
What can I do to address this?
Thanks!
