### 2022-12-05 #4990 @Kevin-Mattheus-Moerman
@aceglia I have changed the title of this issue just now. Next can you please do the following:

- [ ] Update the title in the `paper.md` file
- [ ] Call `@editorialbot generate pdf` in a comment here.

Thanks
