+++ docs/paper.md
@@ -65,7 +65,8 @@ Examples are provided to guide the user and documentation is available.
 
 # Features
 
-`biosiglive` is divided into five independent modules. The main features are described below.  
+`biosiglive` is divided into five independent modules. The main features are:
+
  - `Processing`: real-time and offline data processing.
  - `Interfaces`: interfaces of standard software such as Vicon Nexus (Oxford, UK) or Delsys Trigno Community  (Boston, USA).
  - `Visualization`: real-time signal visualization,
