### 2023-01-10 #4990 @Kevin-Mattheus-Moerman
@alcantarar  @romainmartinez @felixchenier @trallard @abhishektha @melund @modenaxe @BKillen05 @finsberg @bzenger would you be interested in reviewing this submission _Biosiglive: an Open Source Python Package for Real-time Biosignal Processing_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

JOSS reviews focus on the evaluation of [the software](https://github.com/pyomeca/biosiglive) and the review of a [short paper](https://github.com/openjournals/joss-papers/blob/joss.04990/joss.04990/10.21105.joss.04990.pdf).

This is a pre-review issue where reviewers are assigned. Once we have sufficient reviewers we will open a review issue, so the review will take place on GitHub. Here is an example of a complete review if you are curious as to what the process is like:
https://github.com/openjournals/joss-reviews/issues/4859
As you can see, a checklist guides each reviewer through the process.

Let me know if you can help or if you have any more questions. Note we can be quite flexible regarding the time needed to complete the review. Thanks!
