### 2023-01-10 #4990 @felixchenier
Good morning @Kevin-Mattheus-Moerman
Thanks for contacting me. I'm currently on a sabbatical leave, and I do not use Vicon Nexus or Delsys Trigno in my lab, so I couldn't do this review unfortunately.
Regards, Félix.
