+++ biosiglive/processing/data_processing.py
@@ -262,7 +262,7 @@ class GenericProcessing:
             raise RuntimeError("Please choose between low-pass filter and moving average.")
         if low_pass_filter:
             data_proc = self.butter_lowpass_filter(data_proc, self.lpf_lcut, self.data_rate, order=self.lp_butter_order)
-        else:
+        if moving_average:
             w = np.repeat(1, moving_average_window) / moving_average_window
             empty_ma = np.ndarray((data.shape[0], data.shape[1]))
             data_proc = self._moving_average(data_proc, w, empty_ma)
@@ -397,6 +397,7 @@ class RealTimeProcessing(GenericProcessing):
                 moving_average=False,
                 normalization=False,
             )
+            self.processed_data_buffer = emg_proc_tmp / quot
             if low_pass_filter:
                 self.processed_data_buffer = (
                     self.butter_lowpass_filter(emg_proc_tmp, self.lpf_lcut, self.data_rate, order=self.lp_butter_order)
