### 2023-01-21 #4990 @marcoghislieri
Dear @Kevin-Mattheus-Moerman,
Thanks for inviting me to review this submission. Sure, I'm interested and available to review _Biosiglive: an Open Source Python Package for Real-time Biosignal Processing_
All the best
