### 2023-01-22 #5091 @editorialbot
[REVIEW]: Biosiglive: an Open Sources python package for real-time biosignals processing.
**Submitting author:** @aceglia (<a href="http://orcid.org/0000-0002-7854-9410">Amedeo Ceglia</a>)
**Repository:** https://github.com/pyomeca/biosiglive
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.2
