### 2023-01-22 pyomeca/biosiglive#5 @finsberg
[JOSS REVIEW] Fix setup script
I would clean up the [setup script](https://github.com/pyomeca/biosiglive/blob/main/setup.py) so that it is straight forward to install this package without having to read the documentation.

It looks like you have some required dependencies, e.g `matplotlib`, `numpy` , `scipy`. You can read how to specify these in `setup.py` here: https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#declaring-required-dependency.

For the optional dependencies you should also consider to add these to `extras_require`, see https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#optional-dependencies

I would also not pin the dependencies to any exact version (like you do in [requirements.txt](https://github.com/pyomeca/biosiglive/blob/main/requirements.txt)) unless you absolutely need a specific version of a package.
That just makes it hard to combine you library with other libraries. Also, if you specify the dependencies in `setup.py`, then you don't need the `requirements.txt` anymore.

It is also not clear to my why you need to have this line
https://github.com/pyomeca/biosiglive/blob/0aeddc5a02581bb06226296357249f1e014fd76b/setup.py#L11
since I could not find any `.json` files in the repo.

Finally, using a pure `setup.py` file is considered [legacy](https://pip.pypa.io/en/stable/reference/build-system/setup-py/), and you should consider to use a more modern way of specifying the metadata for your project, e.g using [pyproject.toml](https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/)

This is related to https://github.com/openjournals/joss-reviews/issues/5091
