### 2023-01-22 pyomeca/biosiglive#6 @finsberg
[JOSS REVIEW] Add documentation page
The only documentation I could find about this project is within the README file.

I think this project deserves a proper documentation hosted at e.g [GitHub pages](https://pages.github.com) or [Read the docs](https://readthedocs.org). This would make the project much more digestible for more users. Furthermore, you could also host API documentation which you could generate using [Sphinx API doc](https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html) on this page.

On such a documentation page, you should also add separate pages with the examples.

This is related to https://github.com/openjournals/joss-reviews/issues/5091
