### 2023-01-22 pyomeca/biosiglive#7 @finsberg
[JOSS REVIEW] Add `biosiglive` to pypi
I think it is great that you have added `biosiglive` to conda-forge, but for user that don't use conda it would also be great if you could add it to [pypi](https://pypi.org)

This is related to https://github.com/openjournals/joss-reviews/issues/5091
