### 2023-01-22 pyomeca/biosiglive#8 @finsberg
[JOSS REVIEW] Add documentation to each example
I see you have a lot of examples in the [examples folder](https://github.com/pyomeca/biosiglive/tree/main/examples) but it is very hard to know what these examples are doing. It would be good to have proper text explaining each example. One way could be to add each example to a separate folder and add readme file in each folder. Another way could by to embed the documentation directly in the python files. In the latter case I would recommend check out [jupytext](https://jupytext.readthedocs.io) where you can convert the documentation to markdown (or jupyter notebooks), and then you can host the documentation for these examples online (see https://github.com/pyomeca/biosiglive/issues/6).

This is related to https://github.com/openjournals/joss-reviews/issues/5091
