### 2023-01-22 pyomeca/biosiglive#9 @finsberg
[JOSS REVIEW] Add CI to run tests
I see that you have a lot of tests, which is great! However, to make sure that the tests continue to pass you should also set up some continuous integration that runs the tests on every push and pull request to the main branch, see e.g https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python

This is related to https://github.com/openjournals/joss-reviews/issues/5091
