### 2023-01-22 pyomeca/biosiglive#10 @finsberg
[JOSS REVIEW] Add contributing guidelines
As an open source project you should make sure to add some documentation on how to contribute to the project, see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

This is related to https://github.com/openjournals/joss-reviews/issues/5091
