+++ .github/workflows/run_test.yml
@@ -1,0 +1,52 @@
+
+name: Run the tests
+
+on: [pull_request]
+jobs:
+  build:
+    strategy:
+      matrix:
+        include:
+          - os: ubuntu-20.04
+            label: linux-64
+            prefix: /usr/share/miniconda3/envs/biosiglive
+          - os: macos-latest
+            label: osx-64
+            prefix: /Users/runner/miniconda3/envs/biosiglive
+          - os: windows-latest
+            label: win-64
+            prefix: C:\Miniconda3\envs\bioptim
+    name: ${{ matrix.label }}
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash -l {0}
+    steps:
+      - uses: actions/checkout@v2
+      
+      - name: Setup environment
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          miniforge-variant: Mambaforge
+          miniforge-version: latest
+          use-mamba: true
+          activate-environment: biosiglive
+          environment-file: environment.yml
+        
+      - name: Print mamba info
+        run: | 
+          mamba config --show
+          mamba info
+          mamba list
+          
+      - name: Install extra dependencies
+        run: mamba install black pytest-cconda-forge
+
+      - name: Run the actual tests
+        run: pytest -v --color=yes term-missing --cov=biosiglive tests
+          
+      - name: Test installed version of biosiglive
+        run: | 
+          python setup.py install
+          cd
+          python -c "import biosiglive"
