+++ .github/workflows/run_test.yml
@@ -1,7 +1,7 @@
 
 name: Run the tests
 
-on: [pull_request]
+on: [push]
 jobs:
   build:
     strategy:
