+++ biosiglive/environment.yml
@@ -1,0 +1,45 @@
+{% set name = "biosiglive" %}
+{% set version = "2.0.0" %}
+
+package:
+  name: "{{ name|lower }}"
+  version: "{{ version }}"
+
+source:
+  url: https://github.com/pyomeca/biosiglive/releases/download/{{ version }}/biosiglive-{{ version }}.tar.gz
+
+build:
+  number: 0
+  noarch: python
+#  script: "{{ PYTHON }} -m pip install . -vv"
+
+requirements:
+  host:
+    - pip
+    - python >=3.9
+  run:
+    - python >=3.9
+    - matplotlib-base
+    - numpy
+    - scipy
+    - biorbd
+    - bioviz
+    - pyqtgraph
+
+
+test:
+  imports:
+    - biosiglive
+
+about:
+  home: "https://github.com/pyomeca/biosiglive"
+  license: MIT
+  license_family: MIT
+  license_file: LICENSE
+  summary: "A Python package for real-time streaming and processing of biomechanical data"
+  doc_url: https://github.com/pyomeca/biosiglive/blob/master/README.md
+  dev_url: https://github.com/pyomeca/biosiglive
+
+extra:
+  recipe-maintainers:
+    - aceglia
