+++ biosiglive/processing/data_processing.py
@@ -1,5 +1,6 @@
 """
-This file is part of biosiglive. it contains the functions for data processing (offline and in real-time).
+This file contains the functions for data processing (offline and in real-time). Both class herites
+ from the GenericProcessing class.
 """
 
 from scipy.signal import butter, lfilter, filtfilt, convolve
@@ -13,7 +14,7 @@ from ..file_io.save_and_load import save, load
 class GenericProcessing:
     def __init__(self):
         """
-        Initialize the class.
+        Initialize the GenericProcessing class.
         """
         self.bpf_lcut = 10
         self.bpf_hcut = 425
@@ -27,6 +28,7 @@ class GenericProcessing:
     def _butter_bandpass(lowcut: float, highcut: float, fs: float, order: int = 5) -> tuple:
         """
         Create a butter bandpass filter.
+
         Parameters
         ----------
         lowcut : float
@@ -53,6 +55,7 @@ class GenericProcessing:
     def _butter_lowpass(lowcut, fs, order=4) -> tuple:
         """
         Create a butter lowpass filter.
+
         Parameters
         ----------
         lowcut : float
@@ -75,6 +78,7 @@ class GenericProcessing:
     def _butter_bandpass_filter(self, data: np.ndarray, lowcut: float, highcut: float, fs: float, order: int = 5):
         """
         Apply a butter bandpass filter.
+
         Parameters
         ----------
         data: numpy.ndarray
@@ -101,6 +105,7 @@ class GenericProcessing:
     def butter_lowpass_filter(self, data: np.ndarray, lowcut: float, fs: float, order: int = 4) -> np.ndarray:
         """
         Apply a butter lowpass filter.
+
         Parameters
         ----------
         data: numpy.ndarray
@@ -125,12 +130,13 @@ class GenericProcessing:
     def _moving_average(data: np.ndarray, window: np.ndarray, empty_ma: np.ndarray) -> np.ndarray:
         """
         Apply moving average.
+
         Parameters
         ----------
         data: numpy.ndarray
             Data to process.
-        w: numpy.ndarray
-            Window.
+        window: numpy.ndarray
+            window to apply.
         empty_ma: numpy.ndarray
             Empty array to store the moving average.
 
@@ -147,6 +153,7 @@ class GenericProcessing:
     def center(emg_data: np.ndarray, center_value: float = None) -> np.ndarray:
         """
         Center the EMG data.
+
         Parameters
         ----------
         emg_data : numpy.ndarray
@@ -193,6 +200,7 @@ class GenericProcessing:
     def calibration_matrix(self, data: np.ndarray, matrix: np.ndarray) -> np.ndarray:
         """
         Apply a calibration matrix to the data.
+
         Parameters
         ----------
         data: numpy.ndarray
@@ -225,6 +233,7 @@ class GenericProcessing:
     ) -> np.ndarray:
         """
         Process EMG data.
+
         Parameters
         ----------
         data : numpy.ndarray
@@ -275,6 +284,7 @@ class GenericProcessing:
     def update_signal_processing_parameters(self, **kwargs):
         """
         update the signal processing parameters.
+
         Parameters
         ----------
         kwargs: dict
@@ -289,12 +299,13 @@ class RealTimeProcessing(GenericProcessing):
     def __init__(self, data_rate: Union[int, float], processing_window: int = None):
         """
         Initialize the class for real time processing.
+
         Parameters
         ----------
         data_rate : int
             Data rate.
         processing_window : int
-            Processing windows.
+            The window on which data will be processed.
         """
         super().__init__()
         self.data_rate = data_rate
@@ -318,6 +329,7 @@ class RealTimeProcessing(GenericProcessing):
     ) -> np.ndarray:
[…]
