+++ requirements.txt
@@ -1,5 +1,0 @@
-numpy~=1.23.3
-biorbd~=1.9.5
-scipy~=1.9.1
-setuptools~=65.4.0
-matplotlib~=3.6.0
