### 2023-01-25 pyomeca/biosiglive#10 @aceglia
Thanks for this issue @finsberg I have added a "contributing" file in the docs directory and a section in the readme. If you are happy with the PR that was just merged, we can close this issue.
