### 2023-01-25 pyomeca/biosiglive#7 @aceglia
It's done on in the PR #12 that I just merged. Indeed I use conda by default but it is preferable that the package is on both thanks!
I took the opportunity to delete the setup.py and requirements.txt files. So I think we can close this issue and the #5.
