### 2023-01-31 8dc205d
Blacked and update examples (code and docstring)
