+++ examples/EMG_streaming.py
@@ -1,45 +1,30 @@
 """
-This example shows how to stream EMG data using biosiglive and a dedicated interface.
-First an interface is created and then a device is added to the interface. For the EMG data two interfaces are available:
-ViconClient to stream data from a vicon Nexus software through the
-Vicon Data Stream SDK (https://www.vicon.com/software/datastream-sdk/).
-Pytrigno interface to strem data from Delsys trigno community SDK.
-Note that a custom interface is also available from the example 'custom_interface.py' and it allow the user
-to run the examples without any device connection by streaming data from a provided data file.
-If you want to try this example offline you can use the custom interface provided named "MyInterface" and you can use
- it as a standard interface.
+This example shows how to disseminate EMG data using biosiglive and a dedicated interface.
+First an interface is created, then a device is added to the interface. For EMG data, two interfaces are available:
+ViconClient to stream data from a vicon Nexus software via the Vicon Data Stream SDK (https://www.vicon.com/software/datastream-sdk/). The Pytrigno interface for streaming data from the Delsys trigno community SDK (https://delsys.com/sdk/).
+If you want to try this example offline, you can use the provided custom interface named 'MyInterface' which you can use as a standard interface.
 
- The EMG device taking as input argument :
-    - the number of electrodes
-    - the type of the divice (alowwed type are listed in th DeviceType class)
-    - the name of the device (for Vicon system it's need to be the same name as the device to stream).
-    - the rate of the device.
-    - the processing method to use. Here the process_emg method is used.
-    - any other argument needed by the processing method. (you can see the possible argument in the process_emg method
-    documentation).
+The EMG device taking as input argument :
+- the number of electrodes
+- the type of the device (the allowed types are listed in the DeviceType class)
+- the name of the device (for the Vicon system, it must be the same name as the device to stream).
+- the flow of the device.
+- the process method to use. Here, the process_emg method is used.
+- any other argument needed for the processing method. (you can see the possible arguments in the documentation of the process_emg method).
 
-If you want to display the data in real time you can use the biosiglive plot classes.
-You can take a look at the live_plot.py
-After initializing the interface and the device the data streaming take place in a loop.
-First data have to be received from the source though the get_device_data method.
-After that the data might be used as this or can be processed using the process method of the device.
-In this function you pass a method if you want to use a different method than the default one and argument
-for this function as well or to do other processing on the data.
-
-If the display is on the plot.update() method is called with the data to display as argument.
+If you want to display the data in real time, you can use the biosiglive plot classes, please take a look at the live_plot.py example.
+After initializing the interface and the device, the data exchange is done in a loop.
+The data must first be received from the source by the get_device_data method.
+After that, the data can be used as is or processed using the process() method of the Device class.
+In this function you can pass a method if you want to use a different method than the default one and every needed argument for that function as well.
 """
-from custom_interface import MyInterface
+from examples.custom_interface import MyInterface
 from biosiglive.gui.plot import LivePlot
 from biosiglive import (
-    # LivePlot,
-    save,
-    load,
     ViconClient,
     RealTimeProcessingMethod,
-    RealTimeProcessing,
     PlotType,
 )
-
 from time import sleep, time
 
 try:
@@ -48,42 +33,22 @@ except ImportError:
     pass
 
 
[…]
