+++ biosiglive/enums.py
@@ -8,6 +8,7 @@ class InterfaceType(Enum):
     """
     The different types of interfaces that can be used.
     """
+
     ViconClient = "vicon_client"
     PytrignoClient = "pytrigno_client"
     TcpClient = "tcp_client"
@@ -18,6 +19,7 @@ class DeviceType(Enum):
     """
     The different types of devices that can be used.
     """
+
     Emg = "emg"
     Imu = "imu"
     Generic = "generic"
@@ -28,6 +30,7 @@ class MarkerType(Enum):
     """
     The different types of markers that can be used.
     """
+
     Labeled = "labeled"
     Unlabeled = "unlabeled"
 
@@ -36,6 +39,7 @@ class PlotType(Enum):
     """
     The different types of plots that can be used.
     """
+
     Curve = "curve"
     ProgressBar = "progress_bar"
     Skeleton = "skeleton"
@@ -46,6 +50,7 @@ class InverseKinematicsMethods(Enum):
     """
     The different types of inverse kinematics methods that can be used.
     """
+
     BiorbdKalman = "biorbd_kalman"
     BiorbdLeastSquare = "biorbd_least_square"
     Custom = "custom"
@@ -55,6 +60,7 @@ class RealTimeProcessingMethod(Enum):
     """
     The different types of real time processing methods that can be used.
     """
+
     ProcessEmg = "process_emg"
     ProcessImu = "process_imu"
     CalibrationMatrix = "calibration_matrix"
@@ -67,6 +73,7 @@ class OfflineProcessingMethod(Enum):
     """
     The different types of offline processing methods that can be used.
     """
+
     ProcessEmg = "process_emg"
     CalibrationMatrix = "calibration_matrix"
     ComputeMvc = "compute_mvc"

[…]
