+++ docs/source/biosiglive.rst
@@ -1,0 +1,33 @@
+biosiglive package
+==================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   biosiglive.file_io
+   biosiglive.gui
+   biosiglive.interfaces
+   biosiglive.processing
+   biosiglive.streaming
+
+Submodules
+----------
+
+biosiglive.enums module
+-----------------------
+
+.. automodule:: biosiglive.enums
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: biosiglive
+   :members:
+   :undoc-members:
+   :show-inheritance:
