+++ docs/source/biosiglive.streaming.rst
@@ -1,0 +1,45 @@
+biosiglive.streaming package
+============================
+
+Submodules
+----------
+
+biosiglive.streaming.client module
+----------------------------------
+
+.. automodule:: biosiglive.streaming.client
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biosiglive.streaming.server module
+----------------------------------
+
+.. automodule:: biosiglive.streaming.server
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biosiglive.streaming.stream\_data module
+----------------------------------------
+
+.. automodule:: biosiglive.streaming.stream_data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biosiglive.streaming.utils module
+---------------------------------
+
+.. automodule:: biosiglive.streaming.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: biosiglive.streaming
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
