### 2023-01-31 0621dfc
delete auto test to have only docs on this branch
