+++ docs/requirement.txt
@@ -1,3 +1,3 @@
 sphinx
 sphinx_rtd_theme
-myst-parser
+myst_parser
