### 2023-02-01 f209a6e
update run test by ignoring connection test to avoid failed on github
