+++ .github/workflows/run_test.yml
@@ -11,9 +11,9 @@ jobs:
           - os: ubuntu-20.04
             label: linux-64
             prefix: /usr/share/miniconda3/envs/biosiglive
-          - os: macos-latest
-            label: osx-64
-            prefix: /Users/runner/miniconda3/envs/biosiglive
+          #- os: macos-latest
+#            label: osx-64
+#            prefix: /Users/runner/miniconda3/envs/biosiglive
           - os: windows-latest
             label: win-64
             prefix: C:\Miniconda3\envs\biosiglive
