+++ docs/requirements.txt
@@ -1,4 +1,3 @@
 sphinx
 sphinx_rtd_theme
 myst_parser
-biosiglive
