+++ README.md
@@ -1,5 +1,6 @@
 # biosiglive
 `biosiglive` is a python library dedicated to streaming and processing biosignals in real time from Nexus or Delsys Trigno Community.
+Please have a look to the documentation for more information about [biosiglive](https://biosiglive.readthedocs.io/en/latest/index.html). 
 
 # Table of Contents  
 [Testing bioptim](#try-bioptim)
@@ -22,148 +23,21 @@
 
 # How to install
 Biosiglive can be installed from anaconda, pip or from the sources.
-## Installing from anaconda
-You can install biosiglive from anaconda by running the following command :
-```bash
-conda install -c conda-forge biosiglive
-```
-## Installing from pip
-You can install biosiglive from pip by running the following command :
-```bash
-pip install biosiglive
-```
-pyqtgraph, biorbd and bioviz will not be installed in the same time as they are not available on pip or optional.
-So you can install pyqtgraph running the following command :
-```bash
-pip install pyqtgraph
-```
-And you can install biorbd and bioviz running the following command :
-```bash
-conda install -c conda-forge bioviz
-```
+Please refer to the [documentation](https://biosiglive.readthedocs.io/en/latest/install.html) to install biosiglive 
 
-## Installing from the sources
-If installing from the sources, you will need to install the following dependencies from conda (in that particular order):
-- [Python](https://www.python.org/)
-- [matplotlib](https://matplotlib.org/)
-- [numpy](https://numpy.org/)
-- [scipy](https://scipy.org/)
-- [setuptools](https://pypi.org/project/setuptools/)
-- [biorbd](https://github.com/pyomeca/biorbd) (optional: for musculoskeletal models)
-- [pyqtgraph](https://www.pyqtgraph.org/) (optional: for real-time plotting)
-- [pyopengl](https://www.opengl.org/) (optional: for real-time 3D plotting)
-- [bioviz](https://github.com/pyomeca/bioviz) (optional: for skeletal models visualization)
-
-Also, to stream data from Nexus (Vicon) or Trigno Community SDK, you will need to install the following dependencies:
-- [pytrigno](https://github.com/aceglia/pytrigno) (optional: for Trigno Community SDK)
-- [vicon_dssdk](https://www.vicon.com/software/datastream-sdk/) (optional: for Nexus SDK)
-
-After you can install the package by running the following command in the root folder of the package:
-```bash
-pip install .
-```
 # `biosiglive` API
-## GUI
-### LivePlot
-The `LivePlot` class allows the plotting of data in a loop. 
-Data are updated at each ask of the `update` method. It is possible to ask for a plotting rate so that the update is done at a specific frequency.
-Also, it is possible to give to the LivePlot a plot windows, this induces a data buffer that is used to plot the data in a loop.
-Available plot types are listed in the enum `PlotType`:
-- `PlotType.curve`: plot a curve with a given y data. Colors, titles and labels can be given.
-- `PlotType.ProgressBar`: plot a progress bar, the mean value is plotted if several values are given. Name of each bar, units and max values (size of the bar) can be given.
-- `PlotType.Scatter3D`: 3D scatter plot. Colors and size of scatters can be given.
-- `PlotType.Skeleton`: plot a skeleton. Bioviz keys can be given to plot a specific skeleton.
-
-## Interfaces
-### GenericInterface
-The `GenericInterface` class is an abstract class that defines the interface to get data from a system. 
-It is used to define the methods that should be implemented in a specific interface.
-It was used to implement interfaces into `biosiglive` but it can also be used to implement a custom interface (example provided).
-The main methods are:
[…]
