### 2023-02-01 pyomeca/biosiglive#8 @aceglia
@finsberg In fact, examples should be more usefull with a bit more explanations. So I added a dedicated page on the new documentation web site (https://biosiglive.readthedocs.io/en/latest/examples.html). Please let me know if it is what you had in mind and if it suits you.
