### 2023-02-01 pyomeca/biosiglive#6 @aceglia
@finsberg I created a documentation for the library (https://biosiglive.readthedocs.io/). It is a first version and therefore I will update it depends on the users needs. But if it suits you for now we may close this issue. Thanks.
