### 2023-02-01 pyomeca/biosiglive#5 @aceglia
I will close this issue as the setup.py file changed for a pyproject.toml file.
