+++ .github/workflows/run_test.yml
@@ -1,7 +1,7 @@
 
 name: Run the tests
 
-on: [push]
+on: [pull_request]
 
 jobs:
   build:
