### 2023-02-01 da5d97e
Update installation method in environnement
