### 2023-02-01 pyomeca/biosiglive#9 @aceglia
@finsberg I added some github ation workflows. One to check if the code is linted with black and the other to run tests. Obviously tests that needs connection to devices are not run as it is not possible to establish any connection.
The workflow is launched at each PR. If it suites you we can close this issue. Thanks for this issue which will smooth the contribution workflow.
