### 2023-02-02 pyomeca/biosiglive#17 @finsberg
[JOSS REVIEW] Contribution and authorship
The list of authors needs clarification. Three authors are listed in the paper, but only one (or maybe two) author(s) seems to have committed to repository. The command `git shortlog --summary` yields

```
107  Amedeo
1  Pariterre
19  aceglia
72  ceglia
```

Can you please explain the contribution of the other authors?

This is related to https://github.com/openjournals/joss-reviews/issues/5091
