### 2023-02-02 pyomeca/biosiglive#18 @finsberg
[JOSS REVIEW] Problem running `examples/EMG_streaming.py`
When running the `EMG_streaming.py` example I get the following error
```
python EMG_streaming.py
Traceback (most recent call last):
  File "/Users/finsberg/local/src/biosiglive/examples/EMG_streaming.py", line 22, in <module>
    from examples.custom_interface import MyInterface
ModuleNotFoundError: No module named 'examples'
```
Changing the line
```python
from examples.custom_interface import MyInterface
```
to
```python
from custom_interface import MyInterface
```
fixes the problem. Perhaps I am doing something wrong?

This is related to https://github.com/openjournals/joss-reviews/issues/5091
