### 2023-02-02 pyomeca/biosiglive#19 @finsberg
[JOSS REVIEW] Problem running `examples/custom_interface.py`
When running the `custom_interface.py` demo I get the following error

```
custom_interface.py
Traceback (most recent call last):
  File "/Users/finsberg/local/src/biosiglive/examples/custom_interface.py", line 221, in <module>
    print(interface.get_kinematics_from_markers("My markers", get_markers_data=True))
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/finsberg/local/src/biosiglive/examples/custom_interface.py", line 201, in get_kinematics_from_markers
    return self.marker_sets[marker_set_idx].get_kinematics(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/finsberg/mambaforge/envs/biosiglive/lib/python3.11/site-packages/biosiglive/interfaces/param.py", line 279, in get_kinematics
    self.kin_data = self.msk_class.compute_inverse_kinematics(
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/finsberg/mambaforge/envs/biosiglive/lib/python3.11/site-packages/biosiglive/processing/msk_functions.py", line 94, in compute_inverse_kinematics
    ik = biorbd.InverseKinematics(self.model, markers)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/finsberg/mambaforge/envs/biosiglive/lib/python3.11/site-packages/biorbd/rigid_body.py", line 218, in __init__
    raise ValueError(f"The standard dimension of the NumPy array should be (nb_dim, nb_marker, nb_frame)")
ValueError: The standard dimension of the NumPy array should be (nb_dim, nb_marker, nb_frame)
```
I installed `biosiglive` from conda-forge on Mac ARM64.

This is related to https://github.com/openjournals/joss-reviews/issues/5091
