### 2023-02-02 ea4d90e
fix some typo issues in examples and add walk file
