### 2023-02-02 ed8e394
Fix downsamplig for while preparing data for server and update examples and docs
