### 2023-02-02 pyomeca/biosiglive#17 @aceglia
Of course. In the list you gave in the issue, the other contributor is Pariterre but he didn't contribute to this project; he is named here because he is the maintainer of the github group "pyomeca" and he had to accept the project and some first PR before giving me full access.
There are two other authors in my article, one (Felipe Verdugo) who developed the code with me, not in terms of coding but more in developing the methods and testing the code in a real application. The other one (Mickael Begon) is my supervisor and supervised me giving advice about the code, the application or which interfaces to implement, in the whole project and still now.
