### 2023-02-02 pyomeca/biosiglive#18 @aceglia
You did right, the "example." is an artifact in the last PR. I removed this in my last PR and had updated the package on PIP but conda take some time to update. For next runs please be sure to be in version 2.0.2.
