### 2023-02-02 pyomeca/biosiglive#19 @aceglia
Thanks for pointing out the error, it's been a long time since I ran this code, I use the interface directly in the other examples. The problem was that I had to send to "add_markers" a number of markers through the argument nb_markers=xx. I also fixed the problem in my last PR. Sorry for the inconvenience I tested again all examples and there should be no more problems in the version 2.0.2.
