+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 sphinx
 sphinx_rtd_theme
 myst_parser
+numpy
