+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "biosiglive"
-version = "2.0.1"
+version = "2.0.2"
 authors = [
   { name="Amedeo ceglia", email="amedeo.ceglia@umontreal.ca" },
 ]
