+++ docs/requirements.txt
@@ -1,5 +1,3 @@
 sphinx
 sphinx_rtd_theme
 myst_parser
-numpy
-matplotlib
