+++ .readthedocs.yaml
@@ -5,14 +5,14 @@
 # Required
 version: 2
 
-conda:
-    environment: environment.yml
-
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.11"
+    python: "mambaforge-4.10"
+
+conda:
+    environment: environment.yml
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
