+++ .readthedocs.yaml
@@ -12,13 +12,10 @@ build:
     python: "mambaforge-4.10"
 
 conda:
-    environment: environment.yml
+    environment: docs/environment_doc.yml
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/source/conf.py
 
 # Optionally declare the Python requirements required to build your docs
-python:
-   install:
-   - requirements: docs/requirements.txt
