+++ docs/environment_doc.yml
@@ -1,0 +1,19 @@
+name: biosiglive
+channels:
+- conda-forge
+- default
+dependencies:
+- python >=3.10
+- packaging
+- matplotlib-base
+- numpy
+- scipy
+- biorbd
+- bioviz
+- pyqtgraph
+- pip
+- pip:
+  - sphinx
+  - sphinx_rtd_theme
+  - myst_parser
+
