### 2023-02-23 #5091 @marcoghislieri
Dear @Kevin-Mattheus-Moerman,
I've just completed my review of biosiglive python package. Green light for me for publication in JOSS.
All the best
