+++ docs/paper.bib
@@ -6,6 +6,7 @@
   number={66},
   pages={3714},
   year={2021},
+  doi = {https://doi.org/10.21105/joss.03714},
 }
 
 @article{dixon2017biomechzoo,
@@ -16,6 +17,7 @@
   pages={1--10},
   year={2017},
   publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.cmpb.2016.11.007},
 }
 
 @article{martinez2020pyomeca,
@@ -26,16 +28,18 @@
   number={53},
   pages={2431},
   year={2020},
+  doi = {https://doi.org/10.21105/joss.02431},
 }
 
 @inproceedings{Verdugo2022Feeling,
-	author = {Verdugo, Felipe and Ceglia, Amedeo and Frisson, Christian and Burton, Alexandre and Begon, Mickael and Gibet, Sylvie and Wanderley, Marcelo M.},
-	booktitle = {NIME 2022},
-	year = {2022},
-	month = {jun 16},
-	note = {https://nime.pubpub.org/pub/kmn0rbyp},
-	organization = {},
-	title = {Feeling the {Effort} of {Classical} {Musicians} - {A} {Pipeline} from {Electromyography} to {Smartphone} {Vibration} for {Live} {Music} {Performance}},
+  author = {Verdugo, Felipe and Ceglia, Amedeo and Frisson, Christian and Burton, Alexandre and Begon, Mickael and Gibet, Sylvie and Wanderley, Marcelo M.},
+  booktitle = {NIME 2022},
+  year = {2022},
+  month = {jun 16},
+  note = {https://nime.pubpub.org/pub/kmn0rbyp},
+  organization = {},
+  title = {Feeling the {Effort} of {Classical} {Musicians} - {A} {Pipeline} from {Electromyography} to {Smartphone} {Vibration} for {Live} {Music} {Performance}},
+  doi = {https://doi.org/10.21428/92fbeb44.3ce22588},
 }
 
 @article{barre2014biomechanical,
@@ -47,6 +51,7 @@
   pages={80--87},
   year={2014},
   publisher={Elsevier},
+  doi = {10.1016/j.cmpb.2014.01.012},
 }
 
 @article{giggins2013biofeedback,
@@ -58,6 +63,7 @@
   pages={1--11},
   year={2013},
   publisher={Springer},
+  doi = {https://doi.org/10.1186/1743-0003-10-60},
 }
 
 @article{cozean1988biofeedback,
