+++ docs/paper.bib
@@ -6,7 +6,7 @@
   number={66},
   pages={3714},
   year={2021},
-  doi = {https://doi.org/10.21105/joss.03714},
+  doi = {10.21105/joss.03714},
 }
 
 @article{dixon2017biomechzoo,
@@ -17,7 +17,7 @@
   pages={1--10},
   year={2017},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.cmpb.2016.11.007},
+  doi = {10.1016/j.cmpb.2016.11.007},
 }
 
 @article{martinez2020pyomeca,
@@ -28,7 +28,7 @@
   number={53},
   pages={2431},
   year={2020},
-  doi = {https://doi.org/10.21105/joss.02431},
+  doi = {10.21105/joss.02431},
 }
 
 @inproceedings{Verdugo2022Feeling,
@@ -39,7 +39,7 @@
   note = {https://nime.pubpub.org/pub/kmn0rbyp},
   organization = {},
   title = {Feeling the {Effort} of {Classical} {Musicians} - {A} {Pipeline} from {Electromyography} to {Smartphone} {Vibration} for {Live} {Music} {Performance}},
-  doi = {https://doi.org/10.21428/92fbeb44.3ce22588},
+  doi = {10.21428/92fbeb44.3ce22588},
 }
 
 @article{barre2014biomechanical,
@@ -63,7 +63,7 @@
   pages={1--11},
   year={2013},
   publisher={Springer},
-  doi = {https://doi.org/10.1186/1743-0003-10-60},
+  doi = {10.1186/1743-0003-10-60},
 }
 
 @article{cozean1988biofeedback,
