### 2023-03-06 #5091 @aceglia
@Kevin-Mattheus-Moerman

- I fixed the typo in the paper
- The DOI on Zenodo is : 10.5281/zenodo.7703146
- The version is now 2.0.2

Thanks
